CorfuStore: openTable must let caller know if table exists
Callers wish to know if a table exists in the system and never
create it in the process of checking.
To enable such an usecase, they use the openTable() short form
which returns an error of NoSuchElement if that CorfuRuntime
has never ever seen this table.
However this does not tell the caller if the table exists in the
cluster.
So do a stream tails query & find out if the table exists first
and return a different error which will let caller know that the
table they wish to open does exist in the system.